revert weight method to low to high
